Break import cycle at ./internal/utils → ./p2p
