+++ api/service/networkinfo/service.go
@@ -15,6 +15,7 @@ import (
 	nodeconfig "github.com/harmony-one/harmony/internal/configs/node"
 	"github.com/harmony-one/harmony/internal/utils"
 	"github.com/harmony-one/harmony/p2p"
+	p2putils "github.com/harmony-one/harmony/p2p/utils"
 	badger "github.com/ipfs/go-ds-badger"
 	coredis "github.com/libp2p/go-libp2p-core/discovery"
 	libp2pdis "github.com/libp2p/go-libp2p-discovery"
@@ -27,7 +28,7 @@ import (
 type Service struct {
 	Host        p2p.Host
 	Rendezvous  nodeconfig.GroupID
-	bootnodes   utils.AddrList
+	bootnodes   p2putils.AddrList
 	dht         *libp2pdht.IpfsDHT
 	cancel      context.CancelFunc
 	stopChan    chan struct{}
@@ -60,7 +61,7 @@ const (
 // points to a persistent database directory to use.
 func New(
 	h p2p.Host, rendezvous nodeconfig.GroupID, peerChan chan p2p.Peer,
-	bootnodes utils.AddrList, dataStorePath string,
+	bootnodes p2putils.AddrList, dataStorePath string,
 ) (*Service, error) {
 	ctx, cancel := context.WithCancel(context.Background())
 	var dhtOpts []libp2pdhtopts.Option
@@ -98,7 +99,7 @@ func New(
 // MustNew is a panic-on-error version of New.
 func MustNew(
 	h p2p.Host, rendezvous nodeconfig.GroupID, peerChan chan p2p.Peer,
-	bootnodes utils.AddrList, dataStorePath string,
+	bootnodes p2putils.AddrList, dataStorePath string,
 ) *Service {
 	service, err := New(h, rendezvous, peerChan, bootnodes, dataStorePath)
 	if err != nil {
@@ -134,7 +135,7 @@ func (s *Service) Init() error {
 	var wg sync.WaitGroup
 	if s.bootnodes == nil {
 		// TODO: should've passed in bootnodes through constructor.
-		s.bootnodes = utils.BootNodes
+		s.bootnodes = p2putils.BootNodes
 	}
 
 	connected := false
