+++ p2p/utils/flags.go
@@ -1,4 +1,4 @@
-package utils
+package p2putils
 
 import (
 	"fmt"
